backend/requirements.txt
black==21.12b0
Django==4.0
djangorestframework==3.12.4
pytz==2021.3
uWSGI==2.0.20
PyJWT==2.3.0